apps/ai-service/requirements.txt
python-dotenv==1.0.1
numpy>=1.26.0
